jjjj

+++ code/BackEnd_MobileApp/Green_Tech/src/main/java/com/Green_Tech/Green_Tech/Entity/Admin.java
@@ -24,7 +24,7 @@ public class Admin implements UserDetails {
     private String name;
     private String email;
     private String password;
-    private Long phoneNumber;
+    private Integer phoneNumber;
     private Role role;
     private Date createdAt;
     private Date updatedAt;
